refactor: reduce data nesting
